backend/requirement.txt
fastapi>=0.101.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.1
python-multipart>=0.0.6